% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_pdf}
\alias{thesis_pdf}
\title{Create an R Markdown PDF Thesis document}
\usage{
thesis_pdf(template = "template.tex", toc = TRUE, toc_depth = 3,
  toc_bib = TRUE, toc_appendix = TRUE, dev = "cairo_pdf",
  highlight = "default", pandoc_args = NULL, ...)
}
\arguments{
\item{template}{Path of the Pandoc LaTeX template file. Defaults to
\code{"template.tex"}.}

\item{toc}{A Boolean specifying whether table of contents should be created}

\item{toc_depth}{A positive integer}

\item{toc_bib}{Whether to add bibliography to table of contents}

\item{toc_appendix}{Whether to add appendices.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass NULL to prevent syntax highlighting.}

\item{pandoc_args}{optional values for the pandoc_args argument of
bookdown::pdf_book. The argument \code{"--top-level-division=chapter"} is
always included in the pandoc_args.}

\item{...}{other arguments to \code{bookdown::pdf_book()}}

\item{toc_depth}{A positive integer}
}
\value{
A pdf document
}
\description{
Creates a pdf dissertation.
}
